requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
pillow
split-folders
tqdm
python-multipart
huggingface_hub